requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
langchain-anthropic==0.2.4
langchain-community==0.3.5
langchain-core==0.3.15
pydantic==2.9.2
typing-extensions==4.12.2
python-dotenv==1.0.0